Recipe API cleanup and modifications
