+++ app/api/recipes/[id]/route.ts
@@ -1,10 +1,6 @@
 // app/api/recipes/[id]/route.ts
 import { NextResponse } from "next/server";
-import {
-  getRecipeById,
-  updateRecipe,
-  deleteRecipe,
-} from "@/controller/recipeController";
+import { getRecipeById, updateRecipe, deleteRecipe } from "@/controller/recipeController";
 import { uuidSchema } from "@/validation/uuidValidation";
 import { recipeUpdateSchema } from "@/validation/recipeValidation";
 
@@ -16,47 +12,48 @@ import { recipeUpdateSchema } from "@/validation/recipeValidation";
  * @returns JSON with recipe data or an error message.
  */
 export async function GET(
-  req: Request,
-  { params }: { params: Promise<{ id: string }> }
+    req: Request,
+    { params }: { params: Promise<{ id: string }> }
 ): Promise<NextResponse> {
-  let id: string;
+    let id: string;
 
-  // Get ID from path parameters.
-  try {
-    const resolvedParams = await params;
-    id = resolvedParams.id;
-    if (!id) throw new Error();
-  } catch {
-    return NextResponse.json("Not Found", { status: 404 });
-  }
+    // Get ID from path parameters.
+    try 
+    {
+        const resolvedParams = await params;
+        id = resolvedParams.id;
+        if (!id) throw new Error("ID not provided");
+    } 
+    catch 
+    {
+        return NextResponse.json("Not Found", { status: 404 });
+    }
 
-  // Validate that the recipe ID is a UUID.
-  let { error } = await uuidSchema.validate({ uuid: id });
-  if (error) {
-    return NextResponse.json("Bad Request - Invalid Recipe ID", {
-      status: 400,
-    });
-  }
+    // Validate that the recipe ID is a UUID.
+    const { error: idError } = uuidSchema.validate({ uuid: id });
+    if (idError) 
+    {
+        return NextResponse.json("Bad Request - Invalid Recipe ID", { status: 400 });
+    }
 
-  // Get API key from request header and trim it.
-  const rawApiKey = req.headers.get("X-API-Key");
-  const apiKey = rawApiKey ? rawApiKey.trim() : null;
-  if (!apiKey) {
-    return NextResponse.json("Not Authorized", { status: 401 });
-  }
+    // Get API key from request header, trim it, and validate.
+    let apiKey = req.headers.get("X-API-Key");
+    if (!apiKey) 
+    {
+        return NextResponse.json("Not Authorized", { status: 401 });
+    }
+    apiKey = apiKey.trim();
 
-  // Validate that the API key is a UUID.
-  ({ error } = await uuidSchema.validate({ uuid: apiKey }));
-  if (error) {
-    return NextResponse.json("Bad Request - Invalid API Key", {
-      status: 400,
-    });
-  }
+    // Validate the API key.
+    const { error: apiKeyError } = uuidSchema.validate({ uuid: apiKey });
+    if (apiKeyError) 
+    {
+        return NextResponse.json("Bad Request - Invalid API Key", { status: 400 });
+    }
 
-  const recipeResponse = await getRecipeById(id, apiKey);
-  return NextResponse.json(recipeResponse.payload, {
-    status: recipeResponse.status,
-  });
+    // Get the recipe by ID and return it.
+    const recipeResponse = await getRecipeById(id, apiKey);
+    return NextResponse.json(recipeResponse.payload, { status: recipeResponse.status });
 }
 
 /**
@@ -67,47 +64,54 @@ export async function GET(
  * @returns JSON with the updated recipe or an error message.
  */
 export async function PUT(
-  req: Request,
-  { params }: { params: Promise<{ id: string }> }
+    req: Request,
+    { params }: { params: Promise<{ id: string }> }
 ): Promise<NextResponse> {
-  try {
-    const { id } = await params;
-    await uuidSchema.validateAsync({ uuid : id });
+    try 
+    {
+        // Extract and validate the recipe ID.
+        const { id } = await params;
+        const { error: idError } = uuidSchema.validate({ uuid: id });
+        if (idError) 
+        {
+            return NextResponse.json("Bad Request - Invalid Recipe ID", { status: 400 });
+        }
 
-    const apiKey = req.headers.get("X-API-Key");
-    await uuidSchema.validateAsync({ uuid : apiKey });
+        // Retrieve, trim, and validate the API key.
+        let apiKey = req.headers.get("X-API-Key");
+        if (!apiKey) 
+        {
+            return NextResponse.json("Not Authorized", { status: 401 });
+        }
+        apiKey = apiKey.trim();
+        const { error: apiKeyError } = uuidSchema.validate({ uuid: apiKey });
+        if (apiKeyError) 
+        {
+            return NextResponse.json("Bad Request - Invalid API Key", { status: 400 });
+        }
 
-    const body = await req.json();
-    await recipeUpdateSchema.validateAsync(body);
+        // Validate the JSON body against the recipe update schema.
+        const body = await req.json();
+        const { error: bodyError } = recipeUpdateSchema.validate(body);
+        if (bodyError) 
+        {
+            return NextResponse.json({ code: 400, message: bodyError.message }, { status: 400 });
+        }
 
-    const updatedRecipe = await updateRecipe(id, apiKey!, body);
-    if (!updatedRecipe) {
-      return NextResponse.json(
-        { code: 404, message: "Recipe not found" },
-        { status: 404 }
-      );
+        // Update the recipe.
+        const updatedRecipe = await updateRecipe(id, apiKey, body);
+
+        // Return a response based on the update result.
+        if (!updatedRecipe) 
+        {
+            return NextResponse.json({ code: 404, message: "Recipe not found" }, { status: 404 });
+        }
+        return NextResponse.json(updatedRecipe, { status: 200 });
+    } 
+    catch (err: unknown) 
+    {
+        return NextResponse.json({ message: "Server error" }, { status: 500 });
     }
-    return NextResponse.json(updatedRecipe, { status: 200 });
-  } catch (err: unknown) {
-    let code = 500;
-    let message = "Server error";
-    if (
-      typeof err === "object" &&
-      err !== null &&
-      "isJoi" in err &&
-      (err as { isJoi: boolean }).isJoi
-    ) {
-      code = 400;
-      message = (err as unknown as Error).message;
-    } else if (
-      typeof err === "object" &&
-      err !== null &&
-      "message" in err
-    ) {
-      message = (err as unknown as Error).message;
-    }
-    return NextResponse.json({ code, message }, { status: code });
-  }
 }
 
 /**
@@ -118,54 +122,46 @@ export async function PUT(
  * @returns JSON confirming deletion or an error message.
  */
 export async function DELETE(
-  req: Request,
-  { params }: { params: Promise<{ id: string }> }
+    req: Request,
+    { params }: { params: Promise<{ id: string }> }
 ): Promise<NextResponse> {
-  try {
-    // Await the dynamic parameters
-    const { id } = await params;
-    await uuidSchema.validateAsync({ uuid: id });
+    try 
+    {
+        // Extract and validate the recipe ID.
+        const { id } = await params;
+        const { error: idError } = uuidSchema.validate({ uuid: id });
+        if (idError) 
+        {
+            return NextResponse.json("Bad Request - Invalid Recipe ID", { status: 400 });
+        }
 
-    // Get API Key from the request header and trim it
-    const rawApiKey = req.headers.get("X-API-Key");
-    const apiKey = rawApiKey ? rawApiKey.trim() : null;
-    if (!apiKey) {
-      return NextResponse.json("Not Authorized", { status: 401 });
-    }
+        // Retrieve, trim, and validate the API key.
+        let apiKey = req.headers.get("X-API-Key");
+        if (!apiKey) 
+        {
+            return NextResponse.json("Not Authorized", { status: 401 });
+        }
+        apiKey = apiKey.trim();
 
-    // Validate the API Key as a UUID.
-    await uuidSchema.validateAsync({ uuid: apiKey });
+        // Validate the API key.
+        const { error: apiKeyError } = uuidSchema.validate({ uuid: apiKey });
+        if (apiKeyError) 
+        {
+            return NextResponse.json("Bad Request - Invalid API Key", { status: 400 });
+        }
+        
+        // Delete the recipe.
+        const deletionResult = await deleteRecipe(id, apiKey);
 
-    // Call the controller function to delete the recipe.
-    const deletionResult = await deleteRecipe(id, apiKey);
-    if (!deletionResult) {
-      return NextResponse.json(
-        { code: 404, message: "Recipe not found" },
-        { status: 404 }
-      );
+        // Return a response based on the deletion result.
+        if (!deletionResult) 
+        {
+            return NextResponse.json({ code: 404, message: "Recipe not found" }, { status: 404 });
+        }
+        return NextResponse.json({ code: 200, message: "Recipe deleted successfully" }, { status: 200 });
+    } 
+    catch (err: unknown) 
+    {
+        return NextResponse.json({ message: "Server error" }, { status: 500 });
     }
-    return NextResponse.json(
-      { code: 200, message: "Recipe deleted successfully" },
-      { status: 200 }
-    );
-  } catch (err: unknown) {
-    let code = 500;
-    let message = "Server error";
-    if (
-      typeof err === "object" &&
-      err !== null &&
-      "isJoi" in err &&
-      (err as { isJoi: boolean }).isJoi
-    ) {
-      code = 400;
-      message = (err as unknown as Error).message;
-    } else if (
-      typeof err === "object" &&
-      err !== null &&
-      "message" in err
-    ) {
-      message = (err as Error).message;
-    }
-    return NextResponse.json({ code, message }, { status: code });
-  }
 }
